osa4

+++ Osa4/Blogilista/controllers/blogs.js
@@ -1,16 +1,8 @@
 
+const jwt = require('jsonwebtoken')
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
-
-const getTokenFrom = request => {
-  const authorization = request.get('authorization')
-  if (authorization && authorization.toLowerCase().startsWith('bearer ')) {
-    return authorization.substring(7)
-  }
-  return null
-}
-
 
 blogsRouter.get('/blogs', async (request, response) => {
     const blogs = await Blog.find({}).populate('user', {name:1, username:1, id:1})
@@ -22,21 +14,21 @@ blogsRouter.get('/blogs', async (request, response) => {
 blogsRouter.post('/blogs', async (request, response) => {
   console.log(request.body)
   const body = request.body
-  console.log(user)
-  const token = getTokenFrom(request)
+  if (!request.token) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
   const decodedToken = jwt.verify(request.token, process.env.SECRET)
-  if (!token || !decodedToken.id) {
+  if (!request.token || !decodedToken.id) {
     return response.status(401).json({ error: 'token missing or invalid' })
   }
   const user = await User.findById(decodedToken.id)
-
   if(request.body.author && request.body.title){
     if(request.body.likes){
       const blog = new Blog({
-        title:request.body.title,
-        author:request.body.author,
-        url:request.body.url,
-        likes:request.body.likes,
+        title:body.title,
+        author:body.author,
+        url:body.url,
+        likes:body.likes,
         user: user._id
       })
       await blog.save()
@@ -45,15 +37,15 @@ blogsRouter.post('/blogs', async (request, response) => {
       response.status(201).json(blog)
     }else{
       const blog = new Blog({
-        title:request.body.title,
-        author:request.body.author,
-        url:request.body.url,
+        title:body.title,
+        author:body.author,
+        url:body.url,
         likes:0,
         user: user._id
       })
       await blog.save()
       response.status(201).json(blog)
-      user.notes = user.notes.concat(savedNote._id)
+      user.blogs = user.blogs.concat(blog._id)
       await user.save()
     }
   }else{
@@ -62,13 +54,37 @@ blogsRouter.post('/blogs', async (request, response) => {
 })
 
 blogsRouter.delete('/blogs/:id', async (request, response)=> {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+  if (!request.token) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  const blog = await Blog.findById(request.params.id)
+
+  if (!request.token || !decodedToken.id) {
+    console.log('bruh')
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+    const user = await User.findById(decodedToken.id)
+  if(blog.user.toString() === user.id.toString() ){
+    console.log('here2')
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).json(request.body)
+  }else{
+    response.status(401).end()
+  }
+
 })
 
 blogsRouter.put('/blogs/:id', async (request, response)=> {
-  await Blog.findByIdAndUpdate(request.params.id, request.body)
-  response.status(204).json(request.body)
-  
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!request.token || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+    const user = await User.findById(decodedToken.id)
+  if(blog.user.toString() === user.toString() ) {
+    await Blog.findByIdAndUpdate(request.params.id, request.body)
+    response.status(204).json(request.body)
+  }
+
 })
   module.exports = blogsRouter
